+++ backend/main_api/content_client.py
@@ -61,7 +61,7 @@ class A2AContentClientWrapper:
             except Exception as e:
                 self.logger.error(f'获取 AgentCard 失败: {e}', exc_info=True)
                 raise RuntimeError('无法获取 agent card，无法继续运行。') from e
-    async def generate(self, user_question: str, language="English", user_id="") -> None:
+    async def generate(self, user_question: str, metadata={}) -> None:
         """
         user_question: 用户问题
         history： 历史对话消息
@@ -82,7 +82,7 @@ class A2AContentClientWrapper:
                     'role': 'user',
                     'parts': [{'kind': 'text', 'text': user_question}],
                     'messageId': uuid4().hex,
-                    'metadata': {'language': language, "user_id": user_id},
+                    'metadata': metadata,
                     'contextId': self.session_id,
                 },
             }